refactor: [M3-6304] - MUI v5 Migration - `ConditionalWrapper` (#9002)
