--- packages/manager/src/components/ConditionalWrapper/index.ts
@@ -1,2 +1,0 @@
-import ConditionalWrapper from './ConditionalWrapper';
-export default ConditionalWrapper;
