+++ packages/manager/src/components/DocsLink/DocsLink.tsx
@@ -3,7 +3,7 @@ import DocsIcon from 'src/assets/icons/docs.svg';
 import { makeStyles } from 'tss-react/mui';
 import { Theme } from '@mui/material/styles';
 import { sendHelpButtonClickEvent } from 'src/utilities/ga';
-import IconTextLink from '../IconTextLink';
+import { IconTextLink } from '../IconTextLink/IconTextLink';
 
 const useStyles = makeStyles()((theme: Theme) => ({
   root: {
