--- packages/manager/src/components/IconTextLink/index.ts
@@ -1,5 +1,0 @@
-import IconTextLink, { Props as _IconTextLinkProps } from './IconTextLink';
-/* tslint:disable */
-export interface IconTextLinkProps extends _IconTextLinkProps {}
-
-export default IconTextLink;
